// lib/api-client/rq-generated/requests/types.gen.ts
export type OrderDto = {
    id?: string;
    orderDate?: string;
    status?: string;
    orderItems?: Array<OrderItemDto>;
    deliveryFee?: number;
    deliveryAddress?: string;
    totalAmount?: number;
};

export type OrderItemDto = {
    id?: string;
    productVariantId?: string;
    quantity?: number;
    size?: string;
    orderId?: string;
    price?: number;
};